sdk: move shared RPC plumbing to utils.go
